requirements.txt
openai==1.55.3
protobuf==5.29.0
python-dotenv==1.0.1
google-generativeai==0.8.3
